partie printer

+++ app/Http/Controllers/DesignController.php
@@ -40,24 +40,24 @@ class DesignController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate(['design_name'=>'required','version_design'=>'required|in:black,white',
+        $request->validate(['design_name'=>'required|unique:designs,design_name','version_design'=>'required|in:black,white',
         'design_front' => 'required|image|mimes:png,jpg,jpeg|max:80000','design_back' => 'image|mimes:png,jpg,jpeg|max:80000',
         'design_3' => 'image|mimes:png,jpg,jpeg|max:80000','design_4' => 'image|mimes:png,jpg,jpeg|max:80000',]);
-        $design_count=Design::count();
+        // $design_count=Design::count();
         $design=new Design();
-        $design->id=$design_count+1;
+        $design->id_user=Auth::user()->id;
         $design->design_name=$request->design_name;
         $designsNams=["design_front","design_back","design_3","design_4"];
         for($i=0;$i<count($designsNams);$i++){
                 $design_nom=null;
             if(isset($request->{$designsNams[$i]})){
-                $design_nom=$request->design_name."_".$i.".".$request->{$designsNams[$i]}->extension();
+                $design_nom=$request->design_name ."_".$design->id_user. "_" . $i.".".$request->{$designsNams[$i]}->extension();
                 $request->{$designsNams[$i]}->storeAs("images",$design_nom);
             }
             $design->{$designsNams[$i]}=$design_nom;
         }
         $design->version_design=$request->version_design;
-        $design->id_user=Auth::user()->id;
+       
         $design->save();
         return redirect()->route('design.index')->with('success');
     }
@@ -94,30 +94,38 @@ class DesignController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $request->validate(['version_design'=>'in:black,white',
+        'design_front' => 'image|mimes:png,jpg,jpeg|max:80000','design_back' => 'image|mimes:png,jpg,jpeg|max:80000',
+        'design_3' => 'image|mimes:png,jpg,jpeg|max:80000','design_4' => 'image|mimes:png,jpg,jpeg|max:80000',]);
+       
         $design=Design::find($id);
-        $design->design_name=$request->design_name;
-        $designsNams=["design_front","design_back","design_3","design_4"];
-        for ($i = 0; $i < count($designsNams); $i++) {
-            $design_nom = null;
-        
-            if ($request->hasFile($designsNams[$i])) {
-                $design_nom = $request->design_name . "_" . $i . "." . $request->{$designsNams[$i]}->extension();
-                $image_path = public_path('images/' . $design->{$designsNams[$i]});
-                if (file_exists($image_path) && $design->{$designsNams[$i]}!==Null ) {
-                    unlink($image_path);
+       $design_name=Design::where('design_name',$request->design_name)->where('id','!=',$id)->get();
+       if(count($design_name)>=1 ){
+            return redirect()->route('design.index')->with("error","this name of design already existe ");
+       }else{
+            $design->design_name=$request->design_name;
+            $designsNams=["design_front","design_back","design_3","design_4"];
+            for ($i = 0; $i < count($designsNams); $i++) {
+                $design_nom = null;
+                if ($request->hasFile($designsNams[$i])) {
+                    $design_nom = $request->design_name ."_".$design->id_user. "_" . $i . "." . $request->{$designsNams[$i]}->extension();
+                    $image_path = public_path('images/' . $design->{$designsNams[$i]});
+                    if (file_exists($image_path) && $design->{$designsNams[$i]}!==Null ) {
+                        unlink($image_path);
+                    }
+                    $request->{$designsNams[$i]}->storeAs("images", $design_nom);
+                    $design->{$designsNams[$i]} = $design_nom;
+                }else{
+                    $design->{$designsNams[$i]} = $design->{$designsNams[$i]};
                 }
-                $request->{$designsNams[$i]}->storeAs("images", $design_nom);
-                $design->{$designsNams[$i]} = $design_nom;
-            }else{
-                $design->{$designsNams[$i]} = $design->{$designsNams[$i]};
+                
             }
             
+            $design->version_design=$request->version_design;
+            $design->save();
+            return redirect()->route('design.index')->with("succes");
         }
-        
-        $design->version_design=$request->version_design;
-        $design->save();
-        return redirect()->route('design.index')->with("succes");
-    }
+        }
 
     /**
      * Remove the specified resource from storage.
